Running ba test

+++ src/sfm/bundle_adjustment.h
@@ -25,7 +25,7 @@
 class BundleAdjustment {
 
 public:
-    void init(double focal, const cv::Point2d &pp, size_t max_frames);
+    void init(float focal, const cv::Point2d &pp, size_t max_frames);
 
     void addKeyFrame(const VOFrame &frame);
 
@@ -54,7 +54,7 @@ private:
     std::vector<int> pba_2d3d_idx_, pba_cam_idx_;  //index of camera/point for each projection
 
     cv::Point2f pp_;
-    double focal_;
+    float focal_;
 
     size_t max_frames_;
     int count_ = 0;
